+++ CHANGES.md
@@ -4,6 +4,7 @@
 
 ### Unreleased
 
+-   Autofocus on first element in create form; autofocus on edit element
 -   Add onKeyDown function to Input component #58
 -   Add error to Int field overflow #57